Update to 1.19.4

+++ src/main/java/de/maxhenkel/coordfinder/command/CoordCommands.java
@@ -181,7 +181,7 @@ public class CoordCommands {
                     , false);
             return;
         }
-        Location loc = new Location(dimension.dimension().location(), new BlockPos(location.x, location.y, location.z));
+        Location loc = new Location(dimension.dimension().location(), new BlockPos((int) location.x, (int) location.y, (int) location.z));
         CoordFinder.PLACE_CONFIG.setPlace(placeName, loc);
 
         context.getSource().sendSuccess(
